requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2

pydantic-settings~=2.5.2
SQLAlchemy~=2.0.35
redis~=5.1.1
alembic~=1.13.3
python-jose~=3.3.0
httpx~=0.27.2
pytest~=8.3.3
factory_boy~=3.3.1
Faker~=30.8.1
pytest-factoryboy~=2.7.0